+++ examples/rally-api/src/integration-test/groovy/yakworks/rest/BulkRestApiSpec.groovy
@@ -5,7 +5,6 @@ import org.apache.commons.lang3.StringUtils
 import org.springframework.http.HttpStatus
 
 import gorm.tools.job.SyncJobState
-
 import yakworks.rest.client.OkHttpRestTrait
 import grails.testing.mixin.integration.Integration
 import okhttp3.Response
@@ -209,5 +208,17 @@ class BulkRestApiSpec extends Specification implements OkHttpRestTrait {
         failed == 1
         inserted == 1
         updated == 3
+    }
+
+    void "bulk export"() {
+        when:
+        Response resp = get("/api/rally/org/bulk?q=*&parallel=false&async=false")
+        Map body = bodyToMap(resp)
+
+        then:
+        body
+        body.id
+        body.state == SyncJobState.Queued.name()
+        body.sourceId
     }
 }