+++ rally-domain/src/main/groovy/yakworks/rally/mail/model/MailMessage.groovy
@@ -8,11 +8,11 @@ import java.time.LocalDateTime
 
 import groovy.transform.CompileDynamic
 
-import gorm.tools.hibernate.type.JsonType
 import gorm.tools.repository.model.UuidRepoEntity
 import grails.compiler.GrailsCompileStatic
 import grails.persistence.Entity
 import yakworks.commons.transform.IdEqualsHashCode
+import yakworks.gorm.hibernate.type.JsonType
 import yakworks.security.audit.AuditCreatedTrait
 
 import static grails.gorm.hibernate.mapping.MappingBuilder.orm
@@ -21,8 +21,9 @@ import static grails.gorm.hibernate.mapping.MappingBuilder.orm
 @IdEqualsHashCode
 @GrailsCompileStatic
 class MailMessage implements UuidRepoEntity<MailMessage>, AuditCreatedTrait, Serializable {
-    // static belongsTo = [Activity]
+
     UUID id
+
     /** the state of the Mail */
     MsgState state
 
@@ -81,26 +82,26 @@ class MailMessage implements UuidRepoEntity<MailMessage>, AuditCreatedTrait, Ser
     /** The send date, always in Zulu to match server time */
     LocalDateTime sendDate
 
+    // see https://documentation.mailgun.com/docs/mailgun/user-manual/tracking-messages/#webhooks
     @CompileDynamic
     static enum MsgState {
-        Queued, //Queued in the system, ready to be sent to mail processor
-        Sent, //sent from here vai mailgun or smtp
-        Delivered, // if using something like mailgun and we interface, Mailgun sent the email and it was accepted by the recipient email server.
-        Opened, // When using something like mailgun with Open tracking enabled
-        Complained, //The email recipient clicked on the spam complaint button within their email client.
-        Error, //a permanent error occured and message failed, (could be mailgun permanent failure)
-        Failed //Mailgun temporary failure
+        /** Queued in the system, ready to be sent to mail processor */
+        Queued,
+        /** sent from here vai mailgun or smtp */
+        Sent,
+        /** if using something like mailgun and we interface, Mailgun sent the email and it was accepted by the recipient email server. */
+        Delivered,
+        /** When using something like mailgun with Open tracking enabled */
+        Opened,
+        /** The email recipient clicked on the spam complaint button within their email client. */
+        Complained,
+        /** Hard Bounce, (permanent failures) a permanent error occured and message failed, (could be mailgun permanent failure) */
+        Error,
+        /** Soft Bounce, Mailgun Temporary Failure */
+        Failed
     }
 
-
-
-    // static mapping = {
-    //     id generator: 'assigned'
-    // }
-
-
     static mapping = orm {
-        //id(generator: "assigned")
         columns(
             id: property(generator: "assigned"),
             sendTo: property(sqlType: 'TEXT'),
@@ -108,6 +109,7 @@ class MailMessage implements UuidRepoEntity<MailMessage>, AuditCreatedTrait, Ser
             bcc: property(sqlType: 'TEXT'),
             subject: property(sqlType: 'TEXT'),
             body: property(sqlType: 'TEXT'),
+            msgResponse: property(sqlType: 'TEXT'),
             attachmentIds: property(type: JsonType, typeParams: [type: ArrayList]),
             inlineIds: property(type: JsonType, typeParams: [type: ArrayList]),
             tags: property(type: JsonType, typeParams: [type: ArrayList]),
@@ -115,25 +117,24 @@ class MailMessage implements UuidRepoEntity<MailMessage>, AuditCreatedTrait, Ser
     }
 
     static constraintsMap = [
-        sendTo: [
-            d: '''
+        sendTo       : [ d: '''
                 Email address of the recipient(s). RFC standard email format comma seperated with name in quotes and email in < .. >.
-                example: "Blow, Joe" <joe@joes.com>, john@galt.com
-            '''.stripIndent(),
+            ''',
+            example: '"Blow, Joe" <joe@joes.com>, john@galt.com',
             nullable: false, maxSize: 1000
         ],
-        cc:[ d: 'cc addys, see `sendTo` for format', maxSize: 1000],
-        bcc:[ d: 'bcc addys', maxSize: 1000],
-        sendFrom:[ d: 'who its from'],
-        replyTo:[ d: 'what to show for replyTo in email'],
-        source:[ d: 'any special source indicator'],
-        state:[ d: 'the state of the message', nullable: false],
-        subject:[ d: 'email message subject', maxSize: 1000],
-        body:[ d: 'body of message', maxSize: 65535],
-        attachmentIds:[ d: 'ids list of attachments', default: 'plain'],
-        inlineIds:[ d: 'plain, html or markdown for whats in body', default: 'plain'],
-        sendDate: [ d: 'The last response or error message from the mail processor when availiable.' ],
-        msgResponse: [ d: 'The last response or error message from the mail processor when availiable.' ],
-        messageId: [ d: 'When using something like mailgun this is the id returned from the send submit']
+        cc           : [d: 'cc addys, see `sendTo` for format', maxSize: 1000],
+        bcc          : [d: 'bcc addys', maxSize: 1000],
+        sendFrom     : [d: 'who its from'],
+        replyTo      : [d: 'what to show for replyTo in email'],
+        source       : [d: 'any special source indicator'],
+        state        : [d: 'the state of the message', nullable: false],
+        subject      : [d: 'email message subject', maxSize: 1000],
+        body         : [d: 'body of message', maxSize: 65535],
+        attachmentIds: [d: 'ids list of attachments'],
+        inlineIds    : [d: 'Attachment ids with inline disposition.'],
+        sendDate     : [d: 'The send datetime, always in Zulu to match server time'],
+        msgResponse  : [d: 'The last response or error message from the mail processor when availiable.', maxSize: 2000],
+        messageId    : [d: 'When using something like mailgun this is the id returned from the send submit']
     ]
 }